Ninth Commit
30 min - starting to create betting method for the computer player

+++ Pot.java
@@ -5,7 +5,7 @@ public class Pot{
    total = total;
   }
   
-  public void addToPot(int value){
+  public void addToPot(double value){
     total += value;
   }
